stockholm/app/requirements.txt
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.1
pycparser==2.22
python-dotenv==1.0.1
